Make ToggleLayout component using view components
ToggleLayout component contains the page title, button for toggling
views and both grid view and list view components. The different views
with toggling can easily be implemented on any page using this component
