+++ src/components/list-view.js
@@ -7,16 +7,16 @@ import Avatar from '@material-ui/core/Avatar';
 import { GatsbyLink } from "gatsby-theme-material-ui"
 import Img from "gatsby-image"
 
-export default function ListView({ edges }) {
+export default function ListView({ listData }) {
   return (
     <List component="nav">
-        {edges.map(({ node }) => {
+        {listData.map(({ node }) => {
           const { slug } = node.fields
           const { title, image } = node.frontmatter
           return (
             <ListItem button component={GatsbyLink} to={slug} key={node.id}>
               <ListItemAvatar>
-                <Avatar variant="rounded" fixed={image.src.childImageSharp.listImage} component={Img} />
+                <Avatar variant="rounded" fixed={image.src.childImageSharp.listImage} component={Img}/>
               </ListItemAvatar>
               <ListItemText primary={title} secondary={node.excerpt}/>
             </ListItem>
