+++ src/pages/facts.js
@@ -2,52 +2,17 @@ import React from "react"
 import { graphql } from "gatsby"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import GridView from "../components/grid-view"
-import ListView from "../components/list-view"
-import { makeStyles } from "@material-ui/core/styles"
-import Typography from '@material-ui/core/Typography';
-import ViewModuleIcon from '@material-ui/icons/ViewModule';
-import ViewListIcon from '@material-ui/icons/ViewList';
-import { IconButton } from "@material-ui/core"
-import Tooltip from '@material-ui/core/Tooltip';
-import { viewContext } from '../components/view-provider';
-
-
-const useStyles = makeStyles({
-  header: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    margin: 16
-  },
-  toggle: {
-    '@media screen and (max-width: 600px)' : {
-      display: 'none'
-    }
-  }
-});
+import ToggleLayout from '../components/toggle-layout'
 
 const Facts = ({data: { allFacts: { edges }}}) => {
-  const classes = useStyles();
 
   return (
     <Layout>
-      <viewContext.Consumer>
-        {context => (
-          <React.Fragment>
-            <SEO title="All Facts"/>
-            <div className={classes.header}>
-              <Typography variant="h4" component="h1">
-                Facts
-              </Typography>
-              <IconButton className={context.listLayout? (classes.toggle) : (null)} onClick={() => { context.changeLayout() }} >
-                {context.listLayout? (<Tooltip title="Grid View" aria-label="grid view"><ViewModuleIcon/></Tooltip>) : (<Tooltip title="List View" aria-label="list view"><ViewListIcon/></Tooltip>)}
-              </IconButton>
-            </div>
-            {context.listLayout? ( <ListView edges={edges} /> ) : ( <GridView edges={edges} />) }
-          </React.Fragment>
-        )}
-      </viewContext.Consumer>
+      <SEO title="All Facts"/>
+      <ToggleLayout
+      title="Facts"
+      factsData={edges}     
+      />
     </Layout>
   )
 }
